Refactor DicomView
